python/pyproject.toml
[tool.poetry]
name = "science"
version = "0.1.0"
description = ""
authors = ["Jinsoo Kim <jjambbongjoa@duse.co.kr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.19.0"
pymongo = "^4.3.2"
pydantic = {extras = ["email"], version = "^1.10.9"}
python-dotenv = "^0.21.0"
cryptography = "^38.0.3"
validators = "^0.20.0"
python-dateutil = "^2.8.2"
httpx = "^0.23.0"
pybase64 = "^1.2.3"
bcrypt = "^4.0.1"
import-ipynb = "^0.1.4"
dependency-injector = "^4.41.0"
notebook = "^6.5.4"
jupyter = "^1.0.0"
sqlalchemy = "^2.0.15"
pymysql = "^1.0.3"
pytz = "^2023.3"
click = "^8.1.3"
importlib = "^1.0.4"
pydash = "^7.0.4"
numpy = "^1.26.4"
pandas = "^2.2.1"
pytest = "^8.1.1"
platformdirs = "^4.2.0"
fastjsonschema = "^2.19.1"
pydevd-pycharm = "212.5712.39"
fastapi = "0.106.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

